rename error handling script
